use validate-azure-token package instead
