+++ package.json
@@ -36,13 +36,10 @@
     "@types/react": "^17.0.0",
     "@typescript-eslint/eslint-plugin": "^4.22.0",
     "@typescript-eslint/parser": "^4.22.1",
-    "atob": "^2.1.2",
     "axios": "^0.21.1",
     "easy-peasy": "^5.0.0",
-    "jsonwebtoken": "^8.5.1",
     "moment": "^2.29.1",
     "next": "^10.2.3",
-    "node-fetch": "^2.6.1",
     "pixi-viewport": "^4.30.4",
     "pixi.js": "^6.0.2",
     "react": "17.0.2",
@@ -56,7 +53,8 @@
     "sass": "^1.32.11",
     "socket.io": "^4.1.3",
     "socket.io-client": "^4.1.3",
-    "styled-components": "^5.2.2"
+    "styled-components": "^5.2.2",
+    "validate-azure-token": "^1.0.2"
   },
   "devDependencies": {
     "eslint": "^7.17.0",
