+++ pages/api/socket.ts
@@ -2,7 +2,8 @@ import { NextApiRequest } from "next";
 import { NextApiResponseServerIO } from "../../types/next";
 import { Server as ServerIO } from "socket.io";
 import { Server as NetServer } from "http";
-import { validateJWTToken } from "../../utils/validateJWTToken";
+import validateJWTToken from "validate-azure-token";
+import getConfig from "next/config";
 
 export const config = {
   api: {
@@ -23,7 +24,13 @@ const server = (req: NextApiRequest, res: NextApiResponseServerIO) => {
     io.use((socket, next) => {
       // JWT token from the user
       const jwtToken = socket?.handshake?.auth?.token?.split(" ")?.[1]; //Remove the "Bearer" part
-      validateJWTToken(jwtToken)
+      const { serverRuntimeConfig } = getConfig();
+      const verifyOptions = {
+        algorithms: ["RS256"],
+        audience: serverRuntimeConfig.AUDIENCE,
+        issuer: "https://sts.windows.net/3aa4a235-b6e2-48d5-9195-7fcf05b459b0/",
+      };
+      validateJWTToken(jwtToken, verifyOptions)
         .then(() => next())
         .catch((err) => next(new Error(err)));
     });
